restructured CLI and fixed bug in `config::add_files_recursive`
Bug was caused by `common_path::common_path_all` returning the full path when adding a single file.
To fix, the iterator of entry files is checked and if it contains only one file, the parent directory of that file is used as the base instead of the result of `common_path::common_path_all`.